RDKCOM-3626: REFPLTV-1627: RPI 4 64 Bit Build Broken
Reason for change: Fixed build issue for tdk for RPI 4 64 bit hybrid build.
Test Procedure: RPI 4 64 Bit tdk Build should be successful.
Risks: Low

(cherry picked from commit ea61f2b848232de39d498fe8f6940bc6cf8105c2)

+++ DSHal_stub/include/DSHalAgent.h
@@ -49,10 +49,10 @@
 
 using namespace std;
 
-int vpHandle = 0;
-int vdHandle = 0;
-int apHandle = 0;
-int dispHandle = 0;
+intptr_t vpHandle = 0;
+intptr_t vdHandle = 0;
+intptr_t apHandle = 0;
+intptr_t dispHandle = 0;
 string error = "";
 
 class RDKTestAgent;
